Refactored config for Spark modules
